refactor(language_server): move `oxc.fixAll` command logic to `ServerLinter` (#15454)
> This PR refactors the command execution logic by moving it from WorkspaceWorker to ServerLinter. The change consolidates command handling within the linter component, making the worker act as a thin delegation layer.
